split controlTypes.py into submodules (#12510)
The file controlTypes.py is already very complex and long, and the
code can easily be divided into 5 groups.
With the desire to group constants into Enums and maintain backwards
compatibility, which adds length and complexity, it is a canidate to be
split into submodules.

Actions:
- move controlTypes role data and functions to role.py
- move controlTypes state data and functions to state.py
- move isCurrent class to isCurrent.py
- move outputReason to outputReason.py
- move processing states functions to processAndLabelStates.py
