+++ source/controlTypes/role.py
@@ -3,10 +3,8 @@
 # See the file COPYING for more details.
 # Copyright (C) 2007-2021 NV Access Limited, Babbage B.V.
 
-from typing import Dict, Union, Set, Any, Optional, List
-from enum import Enum, auto
+from typing import Dict
 
-from logHandler import log
 
 ROLE_UNKNOWN=0
 ROLE_WINDOW=1
@@ -160,50 +158,6 @@ ROLE_REGION = 151
 ROLE_FIGURE = 152
 ROLE_MARKED_CONTENT = 153
 
-STATE_UNAVAILABLE=0X1
-STATE_FOCUSED=0X2
-STATE_SELECTED=0X4
-STATE_BUSY=0X8
-STATE_PRESSED=0X10
-STATE_CHECKED=0X20
-STATE_HALFCHECKED=0X40
-STATE_READONLY=0X80
-STATE_EXPANDED=0X100
-STATE_COLLAPSED=0X200
-STATE_INVISIBLE=0X400
-STATE_VISITED=0X800
-STATE_LINKED=0X1000
-STATE_HASPOPUP=0X2000
-STATE_PROTECTED=0X4000
-STATE_REQUIRED=0X8000
-STATE_DEFUNCT=0X10000
-STATE_INVALID_ENTRY=0X20000
-STATE_MODAL=0X40000
-STATE_AUTOCOMPLETE=0x80000
-STATE_MULTILINE=0X100000
-STATE_ICONIFIED=0x200000
-STATE_OFFSCREEN=0x400000
-STATE_SELECTABLE=0x800000
-STATE_FOCUSABLE=0x1000000
-STATE_CLICKABLE=0x2000000
-STATE_EDITABLE=0x4000000
-STATE_CHECKABLE=0x8000000
-STATE_DRAGGABLE=0x10000000
-STATE_DRAGGING=0x20000000
-STATE_DROPTARGET=0x40000000
-STATE_SORTED=0x80000000
-STATE_SORTED_ASCENDING=0x100000000
-STATE_SORTED_DESCENDING=0x200000000
-STATES_SORTED=frozenset([STATE_SORTED,STATE_SORTED_ASCENDING,STATE_SORTED_DESCENDING])
-STATE_HASLONGDESC=0x400000000
-STATE_PINNED=0x800000000
-STATE_HASFORMULA=0x1000000000 #Mostly for spreadsheets
-STATE_HASCOMMENT=0X2000000000
-STATE_OBSCURED=0x4000000000
-STATE_CROPPED=0x8000000000
-STATE_OVERFLOWING=0x10000000000
-STATE_UNLOCKED=0x20000000000
-STATE_HAS_ARIA_DETAILS = 0x40000000000
 
 roleLabels: Dict[int, str] = {
 	# Translators: The word for an unknown control type.
@@ -510,99 +464,6 @@ roleLabels: Dict[int, str] = {
 	ROLE_MARKED_CONTENT: _("marked content"),
 }
 
-stateLabels: Dict[int, str] = {
-	# Translators: This is presented when a control or document is unavailable.
-	STATE_UNAVAILABLE:_("unavailable"),
-	# Translators: This is presented when a control has focus.
-	STATE_FOCUSED:_("focused"),
-	# Translators: This is presented when the control is selected.
-	STATE_SELECTED:_("selected"),
-	# Translators: This is presented when a document is busy.
-	STATE_BUSY:_("busy"),
-	# Translators: This is presented when a button is pressed.
-	STATE_PRESSED:_("pressed"),
-	# Translators: This is presented when a check box is checked.
-	STATE_CHECKED:_("checked"),
-	# Translators: This is presented when a three state check box is half checked.
-	STATE_HALFCHECKED:_("half checked"),
-	# Translators: This is presented when the control is a read-only control such as read-only edit box.
-	STATE_READONLY:_("read only"),
-	# Translators: This is presented when a tree view or submenu item is expanded.
-	STATE_EXPANDED:_("expanded"),
-	# Translators: This is presented when a tree view or submenu is collapsed.
-	STATE_COLLAPSED:_("collapsed"),
-	# Translators: This is presented when a control or a document becomes invisible.
-	STATE_INVISIBLE:_("invisible"),
-	# Translators: This is presented when a visited link is encountered.
-	STATE_VISITED:_("visited"),
-	# Translators: This is presented when a link is encountered.
-	STATE_LINKED:_("linked"),
-	# Translators: This is presented when the control menu item has a submenu.
-	STATE_HASPOPUP:_("subMenu"),
-	# Translators: This is presented when a protected control or a document is encountered.
-	STATE_PROTECTED:_("protected"),
-	# Translators: This is presented when a required form field is encountered.
-	STATE_REQUIRED:_("required"),
-	# Translators: Reported when an object no longer exists in the user interface;
-	# i.e. it is dead and is no longer usable.
-	STATE_DEFUNCT:_("defunct"),
-	# Translators: This is presented when an invalid entry has been made.
-	STATE_INVALID_ENTRY:_("invalid entry"),
-	STATE_MODAL:_("modal"),
-	# Translators: This is presented when a field supports auto completion of entered text such as email address field in Microsoft Outlook.
-	STATE_AUTOCOMPLETE:_("has auto complete"),
-	# Translators: This is presented when an edit field allows typing multiple lines of text such as comment fields on websites.
-	STATE_MULTILINE:_("multi line"),
-	STATE_ICONIFIED:_("iconified"),
-	# Translators: Presented when the current control is located off screen.
-	STATE_OFFSCREEN:_("off screen"),
-	# Translators: Presented when the control allows selection such as text fields.
-	STATE_SELECTABLE:_("selectable"),
-	# Translators: Presented when a control can be moved to using system focus.
-	STATE_FOCUSABLE:_("focusable"),
-	# Translators: Presented when a control allows clicking via mouse (mostly presented on web controls).
-	STATE_CLICKABLE:_("clickable"),
-	STATE_EDITABLE:_("editable"),
-	STATE_CHECKABLE:_("checkable"),
-	STATE_DRAGGABLE:_("draggable"),
-	STATE_DRAGGING:_("dragging"),
-	# Translators: Reported where an object which is being dragged can be dropped.
-	# This is only reported for objects that support accessible drag and drop.
-	STATE_DROPTARGET:_("drop target"),
-	STATE_SORTED:_("sorted"),
-	STATE_SORTED_ASCENDING:_("sorted ascending"),
-	STATE_SORTED_DESCENDING:_("sorted descending"),
-	# Translators: a state that denotes that an object (usually a graphic) has a long description.
-	STATE_HASLONGDESC:_("has long description"),
-	# Translators: a state that denotes that an object has additional details (such as a comment section).
-	STATE_HAS_ARIA_DETAILS: _("has details"),
-	# Translators: a state that denotes that an object is pinned in its current location
-	STATE_PINNED:_("pinned"),
-	# Translators: a state that denotes the existance of a formula on a spreadsheet cell
-	STATE_HASFORMULA:_("has formula"),
-	# Translators: a state that denotes the existance of a comment.
-	STATE_HASCOMMENT:_("has comment"),
-	# Translators: a state that denotes that the object is covered partially or fully by another object
-	STATE_OBSCURED:_("obscured"),
-	# Translators: a state that denotes that the object(text) is cropped as it couldn't be accommodated in the allocated/available space
-	STATE_CROPPED:_("cropped"),
-	# Translators: a state that denotes that the object(text) is overflowing into the adjacent space
-	STATE_OVERFLOWING:_("overflowing"),
-	# Translators: a state that denotes that the object is unlocked (such as an unlocked cell in a protected Excel spreadsheet). 
-	STATE_UNLOCKED:_("unlocked"),
-}
-
-negativeStateLabels={
-	# Translators: This is presented when a selectable object (e.g. a list item) is not selected.
-	STATE_SELECTED:_("not selected"),
-	# Translators: This is presented when a button is not pressed.
-	STATE_PRESSED:_("not pressed"),
-	# Translators: This is presented when a checkbox is not checked.
-	STATE_CHECKED:_("not checked"),
-	# Translators: This is presented when drag and drop is finished.
-	# This is only reported for objects which support accessible drag and drop.
-	STATE_DROPTARGET:_("done dragging"),
-}
 
 silentRolesOnFocus={
 	ROLE_PANE,
@@ -619,6 +480,7 @@ silentRolesOnFocus={
 	ROLE_BORDER,
 }
 
+
 silentValuesForRoles={
 	ROLE_CHECKBOX,
 	ROLE_RADIOBUTTON,
@@ -626,244 +488,3 @@ silentValuesForRoles={
 	ROLE_MENUITEM,
 	ROLE_APPLICATION,
 }
-
-
-class OutputReason(Enum):
-	"""Specify the reason that a given piece of output was generated.
-	"""
-	#: An object to be reported due to a focus change or similar.
-	FOCUS = auto()
-	#: An ancestor of the focus object to be reported due to a focus change or similar.
-	FOCUSENTERED = auto()
-	#: An item under the mouse.
-	MOUSE = auto()
-	#: A response to a user query.
-	QUERY = auto()
-	#: Reporting a change to an object.
-	CHANGE = auto()
-	#: A generic, screen reader specific message.
-	MESSAGE = auto()
-	#: Text reported as part of a say all.
-	SAYALL = auto()
-	#: Content reported due to caret movement or similar.
-	CARET = auto()
-	#: No output, but any state should be cached as if output had occurred.
-	ONLYCACHE = auto()
-
-	QUICKNAV = auto()
-
-
-class IsCurrent(Enum):
-	"""Values to use within NVDA to denote 'current' values.
-	These describe if an item is the current item within a particular kind of selection.
-	EG aria-current
-	"""
-	NO = "false"
-	YES = "true"
-	PAGE = "page"
-	STEP = "step"
-	LOCATION = "location"
-	DATE = "date"
-	TIME = "time"
-
-	@property
-	def displayString(self):
-		"""
-		@return: The translated UI display string that should be used for this value of the IsCurrent enum
-		"""
-		try:
-			return _isCurrentLabels[self]
-		except KeyError:
-			log.debugWarning(f"No translation mapping for: {self}")
-			# there is a value for 'current' but NVDA hasn't learned about it yet,
-			# at least describe in the general sense that this item is 'current'
-			return _isCurrentLabels[IsCurrent.YES]
-
-
-#: Text to use for 'current' values. These describe if an item is the current item
-#: within a particular kind of selection. EG aria-current
-_isCurrentLabels: Dict[Enum, str] = {
-	IsCurrent.NO: "",  # There is nothing extra to say for items that are not current.
-	# Translators: Presented when an item is marked as current in a collection of items
-	IsCurrent.YES: _("current"),
-	# Translators: Presented when a page item is marked as current in a collection of page items
-	IsCurrent.PAGE: _("current page"),
-	# Translators: Presented when a step item is marked as current in a collection of step items
-	IsCurrent.STEP: _("current step"),
-	# Translators: Presented when a location item is marked as current in a collection of location items
-	IsCurrent.LOCATION: _("current location"),
-	# Translators: Presented when a date item is marked as current in a collection of date items
-	IsCurrent.DATE: _("current date"),
-	# Translators: Presented when a time item is marked as current in a collection of time items
-	IsCurrent.TIME: _("current time"),
-}
-
-
-def processPositiveStates(role, states, reason: OutputReason, positiveStates=None):
-	"""Processes the states for an object and returns the positive states to output for a specified reason.
-	For example, if C{STATE_CHECKED} is in the returned states, it means that the processed object is checked.
-	@param role: The role of the object to process states for (e.g. C{ROLE_CHECKBOX}.
-	@type role: int
-	@param states: The raw states for an object to process.
-	@type states: set
-	@param reason: The reason to process the states (e.g. C{OutputReason.FOCUS}.
-	@param positiveStates: Used for C{OutputReason.CHANGE}, specifies states changed from negative to positive;
-	@type positiveStates: set
-	@return: The processed positive states.
-	@rtype: set
-	"""
-	positiveStates = positiveStates.copy() if positiveStates is not None else states.copy()
-	# The user never cares about certain states.
-	if role==ROLE_EDITABLETEXT:
-		positiveStates.discard(STATE_EDITABLE)
-	if role!=ROLE_LINK:
-		positiveStates.discard(STATE_VISITED)
-	positiveStates.discard(STATE_SELECTABLE)
-	positiveStates.discard(STATE_FOCUSABLE)
-	positiveStates.discard(STATE_CHECKABLE)
-	if STATE_DRAGGING in positiveStates:
-		# It's obvious that the control is draggable if it's being dragged.
-		positiveStates.discard(STATE_DRAGGABLE)
-	if role == ROLE_COMBOBOX:
-		# Combo boxes inherently have a popup, so don't report it.
-		positiveStates.discard(STATE_HASPOPUP)
-	import config
-	if not config.conf['documentFormatting']['reportClickable'] or role in (ROLE_LINK, ROLE_BUTTON, ROLE_CHECKBOX, ROLE_RADIOBUTTON, ROLE_TOGGLEBUTTON, ROLE_MENUITEM, ROLE_TAB, ROLE_SLIDER, ROLE_DOCUMENT, ROLE_CHECKMENUITEM, ROLE_RADIOMENUITEM):
-		# This control is clearly clickable according to its role,
-		# or reporting clickable just isn't useful,
-		# or the user has explicitly requested no reporting clickable
-		positiveStates.discard(STATE_CLICKABLE)
-	if reason == OutputReason.QUERY:
-		return positiveStates
-	positiveStates.discard(STATE_DEFUNCT)
-	positiveStates.discard(STATE_MODAL)
-	positiveStates.discard(STATE_FOCUSED)
-	positiveStates.discard(STATE_OFFSCREEN)
-	positiveStates.discard(STATE_INVISIBLE)
-	if reason != OutputReason.CHANGE:
-		positiveStates.discard(STATE_LINKED)
-		if role in (
-			ROLE_LISTITEM,
-			ROLE_TREEVIEWITEM,
-			ROLE_MENUITEM,
-			ROLE_TABLEROW,
-			ROLE_CHECKBOX,
-		) and STATE_SELECTABLE in states:
-			positiveStates.discard(STATE_SELECTED)
-	if role not in (ROLE_EDITABLETEXT, ROLE_CHECKBOX):
-		positiveStates.discard(STATE_READONLY)
-	if role == ROLE_CHECKBOX:
-		positiveStates.discard(STATE_PRESSED)
-	if role == ROLE_MENUITEM and STATE_HASPOPUP in positiveStates:
-		# The user doesn't usually care if a submenu is expanded or collapsed.
-		positiveStates.discard(STATE_COLLAPSED)
-		positiveStates.discard(STATE_EXPANDED)
-	if STATE_FOCUSABLE not in states:
-		positiveStates.discard(STATE_EDITABLE)
-	if not config.conf["annotations"]["reportDetails"]:
-		# reading aria-details is an experimental feature still and should not always be reported.
-		positiveStates.discard(STATE_HAS_ARIA_DETAILS)
-	return positiveStates
-
-
-def processNegativeStates(role, states, reason: OutputReason, negativeStates=None):
-	"""Processes the states for an object and returns the negative states to output for a specified reason.
-	For example, if C{STATE_CHECKED} is in the returned states, it means that the processed object is not checked.
-	@param role: The role of the object to process states for (e.g. C{ROLE_CHECKBOX}.
-	@type role: int
-	@param states: The raw states for an object to process.
-	@type states: set
-	@param reason: The reason to process the states (e.g. C{OutputReason.FOCUS}.
-	@param negativeStates: Used for C{OutputReason.CHANGE}, specifies states changed from positive to negative;
-	@type negativeStates: set
-	@return: The processed negative states.
-	@rtype: set
-	"""
-	if reason == OutputReason.CHANGE and not isinstance(negativeStates, set):
-		raise TypeError("negativeStates must be a set for this reason")
-	speakNegatives = set()
-	# Add the negative selected state if the control is selectable,
-	# but only if it is reported for the reason of focus, or this is a change to the focused object. 
-	# The condition stops "not selected" from being spoken in some broken controls
-	# when the state change for the previous focus is issued before the focus change.
-	if (
-		# Only include if the object is actually selectable
-		STATE_SELECTABLE in states
-		# Only include if the object is focusable (E.g. ARIA grid cells, but not standard html tables)
-		and STATE_FOCUSABLE in states
-		# Only include  if reporting the focus or when states are changing on the focus.
-		# This is to avoid exposing it for things like caret movement in browse mode. 
-		and (reason == OutputReason.FOCUS or (reason == OutputReason.CHANGE and STATE_FOCUSED in states))
-		and role in (
-			ROLE_LISTITEM, 
-			ROLE_TREEVIEWITEM, 
-			ROLE_TABLEROW,
-			ROLE_TABLECELL,
-			ROLE_TABLECOLUMNHEADER,
-			ROLE_TABLEROWHEADER,
-			ROLE_CHECKBOX,
-		)
-	):
-		speakNegatives.add(STATE_SELECTED)
-	# Restrict "not checked" in a similar way to "not selected".
-	if(
-		(role in (ROLE_CHECKBOX, ROLE_RADIOBUTTON, ROLE_CHECKMENUITEM) or STATE_CHECKABLE in states)
-		and (STATE_HALFCHECKED not in states)
-		and (reason != OutputReason.CHANGE or STATE_FOCUSED in states)
-	):
-		speakNegatives.add(STATE_CHECKED)
-	if role == ROLE_TOGGLEBUTTON:
-		speakNegatives.add(STATE_PRESSED)
-	if reason == OutputReason.CHANGE:
-		# We want to speak this state only if it is changing to negative.
-		speakNegatives.add(STATE_DROPTARGET)
-		# We were given states which have changed to negative.
-		# Return only those supplied negative states which should be spoken;
-		# i.e. the states in both sets.
-		speakNegatives &= negativeStates
-		# #6946: if HALFCHECKED is present but CHECKED isn't, we should make sure we add CHECKED to speakNegatives.
-		if (STATE_HALFCHECKED in negativeStates and STATE_CHECKED not in states):
-			speakNegatives.add(STATE_CHECKED)
-		if STATES_SORTED & negativeStates and not STATES_SORTED & states:
-			# If the object has just stopped being sorted, just report not sorted.
-			# The user doesn't care how it was sorted before.
-			speakNegatives.add(STATE_SORTED)
-		return speakNegatives
-	else:
-		# This is not a state change; only positive states were supplied.
-		# Return all negative states which should be spoken, excluding the positive states.
-		return speakNegatives - states
-
-
-def processAndLabelStates(
-		role: int,
-		states: Set[Any],
-		reason: OutputReason,
-		positiveStates: Optional[Set[Any]] = None,
-		negativeStates: Optional[Set[Any]] = None,
-		positiveStateLabelDict: Dict[int, str] = {},
-		negativeStateLabelDict: Dict[int, str] = {},
-) -> List[str]:
-	"""Processes the states for an object and returns the appropriate state labels for both positive and negative states.
-	@param role: The role of the object to process states for (e.g. C{ROLE_CHECKBOX}.
-	@param states: The raw states for an object to process.
-	@param reason: The reason to process the states (e.g. C{OutputReason.FOCUS}.
-	@param positiveStates: Used for C{OutputReason.CHANGE}, specifies states changed from negative to positive;
-	@param negativeStates: Used for C{OutputReason.CHANGE}, specifies states changed from positive to negative;
-	@param positiveStateLabelDict: Dictionary containing state identifiers as keys and associated positive labels as their values.
-	@param negativeStateLabelDict: Dictionary containing state identifiers as keys and associated negative labels as their values.
-	@return: The labels of the relevant positive and negative states.
-	"""
-	mergedStateLabels=[]
-	positiveStates = processPositiveStates(role, states, reason, positiveStates)
-	negativeStates = processNegativeStates(role, states, reason, negativeStates)
-	for state in sorted(positiveStates | negativeStates):
-		if state in positiveStates:
-			mergedStateLabels.append(positiveStateLabelDict.get(state, stateLabels[state]))
-		elif state in negativeStates:
-			# Translators: Indicates that a particular state of an object is negated.
-			# Separate strings have now been defined for commonly negated states (e.g. not selected and not checked),
-			# but this still might be used in some other cases.
-			# %s will be replaced with the full identifier of the negated state (e.g. selected).
-			mergedStateLabels.append(negativeStateLabelDict.get(state, negativeStateLabels.get(state, _("not %s") % stateLabels[state])))
-	return mergedStateLabels
